Corrigindo verificação de usuário para retornar mensagens

+++ app/Http/Controllers/User/AdContactController.php
@@ -52,7 +52,7 @@ class AdContactController extends Controller
      */
     public function show(Ad $ad, Contact $contact)
     {
-        if ($ad->user_id !== request()->user()->id) {
+        if ((int) $ad->user_id !== (int) request()->user()->id) {
             return response()->json(null, 403);
         }
         if (empty($contact->viewed_at)) {
@@ -71,7 +71,7 @@ class AdContactController extends Controller
      */
     public function destroy(Ad $ad, Contact $contact)
     {
-        if ($ad->user_id !== request()->user()->id) {
+        if ((int) $ad->user_id !== (int) request()->user()->id) {
             return response()->json(null, 403);
         }
         if ($contact->delete()) {
